requirements.txt
streamlit
pandas
numpy
mysql-connector-python
statsmodels
plotly
scikit-learn
